requirements.txt
#aimrocks==0.2.0
aiodns==3.0.0
aleph_alpha_client==3.1.0
apify_client==1.1.1
arize==7.0.1
asynctest==0.13.0
atlassian==0.0.0
azure_storage==0.37.0
banana_dev==4.0.2
beautifulsoup4==4.12.2
bilibili_api==9.1.0
boto3==1.26.131
botocore==1.29.131
Brotli==1.0.9
brotlicffi==1.0.9.2
cchardet==2.1.7
#cerebrium==0.4.34
chromadb==0.3.22
clearml==1.10.4
clickhouse_connect==0.5.23
comet_ml==3.33.1
ConfigParser==5.3.0
contextlib2==21.6.0
cryptography==40.0.2
cx_Oracle==8.3.0
Cython==0.29.33
datasets==2.12.0
disco==1.40.4
dl==0.1.0
docarray==0.31.1
docutils==0.20
docx2txt==0.8
duckdb==0.7.1
elasticsearch==8.7.0
email_validator==2.0.0.post2
extract_msg==0.41.0
fake_useragent==1.1.3
firebase_admin==6.1.0
fitz==0.0.1.dev2
genapi==0.0.8
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
googlemaps==4.10.0
gptcache==0.1.22
greenlet==2.0.2
guardrails==2.0.0
gunicorn==20.1.0
HTMLParser==0.0.2
hypothesis==6.75.2
idna_ssl==1.1.0
importlib_metadata==5.0.0
InstructorEmbedding==1.0.0
ipython==8.13.2
ipywidgets==7.7.0
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
lark==1.1.5
lockfile==0.12.2
lxml==4.9.2
matplotlib==3.6.2
metal_sdk==1.0.1
mtrand==0.1
mwparserfromhell==0.6.4
mwxml==0.3.3
networkx==2.8.8
nltk==3.8.1
nomic==1.1.6
nose==1.3.7
numarray==1.5.1
Numeric==24.2
O365==2.0.26
ordereddict==1.1
pdf2image==1.16.3
pdfminer==20191125
pexpect==4.8.0
pickle5==0.0.11
Pillow==9.5.0
pinecone_text==0.4.2
pipeline==0.1
praw==7.7.0
promptlayer==0.1.80
protobuf==4.23.0
psutil==5.9.1
pygpt4all==1.1.0
PyInstaller==5.10.1
pymilvus==2.2.8
pyOpenSSL==23.1.1
pyowm==3.3.0
pypandoc==1.11
pypdf==3.8.1
PyPDF2==3.0.1
pyspark==3.4.0
pysqlcipher3==1.2.0
pysrt==1.1.2
pytesseract==0.3.10
pytest==7.3.1
python_magic==0.4.27
pytube==15.0.0
pytz==2022.6
qdrant_client==1.1.7
railroad==0.5.0
reportlab==4.0.0
runhouse==0.0.5
scikit_learn==1.2.2
scipy==1.10.1
selenium==4.9.1
sentence_transformers==2.2.2
setuptools_scm==7.1.0
simplejson==3.19.1
slack_sdk==3.21.3
spacy==3.5.2
Sphinx==7.0.0
sqlcipher3==0.5.0
svglib==1.5.1
tensorflow==2.12.0
tensorflow_text==2.12.1
textstat==0.7.3
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tokio==0.2.0
torch==2.0.1
tornado==6.1
transformers==4.28.1
truststore==0.7.0
tweepy==4.14.0
urllib3_secure_extra==0.1.0
uvloop==0.17.0
vespa==0.6
wandb==0.15.2
wincertstore==0.2.1
wolframalpha==5.0.0
xlrd==2.0.1
xmlrpclib==1.0.1
youtube_transcript_api==0.6.0
zstandard==0.21.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.26
langchain-groq==0.1.9
python-dotenv==1.0.1
streamlit==1.37.0
